# go-authz/authorization/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Authorization API documentation.",
        "title": "Authorization API",
        "contact": {
            "name": "Jamessaldo",
            "url": "https://github.com/jamessaldo/collections/issues",
            "email": "ghozyghlmlaff@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api/v1",
    "paths": {
        "/invitations/verify": {
            "post": {
                "description": "Get invitation data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Get invitation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/invitations/{id}": {
            "get": {
                "description": "Get invitation data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Get invitation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InvitationRetreivalSchema"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete invitation data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Delete invitation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/teams/{id}/invitation/{invitation_id}": {
            "post": {
                "description": "Resend invitation to join team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Resend Invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Invitation ID",
                        "name": "invitation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.InvitationRetreivalSchema": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sender_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "team": {
                    "$ref": "#/definitions/dto.TeamRetrievalSchema"
                }
            }
        },
        "dto.MembershipRetrievalSchema": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user": {}
            }
        },
...
        "dto.TeamRetrievalSchema": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator": {},
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_personal": {
                    "type": "boolean"
                },
                "last_active_at": {
                    "type": "string"
                },
                "memberships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MembershipRetrievalSchema"
                    }
                },
                "name": {
                    "type": "string"
                },
                "num_of_members": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}